Move lock and helper functions out of lib.rs for clarity
